+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
+import Expenses from './components/Feature/Expenses/Expenses';
 
-import Expenses from "./components/Expenses/Expenses";
 
 function App() {
   const expenses = [
@@ -25,18 +25,18 @@ function App() {
     },
   ];
 
-  return React.createElement(
-    'div',
-    {},
-    React.createElement('h2', {}, "Let's get started!"),
-    React.createElement(Expenses, { items: expenses }),
-  );
-  // return (
-  //   <div>
-  //     <h2>Let's get started!</h2>
-  //     <Expenses items={expenses}/>
-  //   </div>
+  // return React.createElement(
+  //   'div',
+  //   {},
+  //   React.createElement('h2', {}, "Let's get started!"),
+  //   React.createElement(Expenses, { items: expenses }),
   // );
+  return (
+    <div>
+      <h2>Let's get started!</h2>
+      <Expenses items={expenses}/>
+    </div>
+  );
 }
 
 export default App;
