+++ src/components/Feature/ExpenseItem/ExpenseItem.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Card from "../Card/Card";
+import Card from "../../UI/Card/Card";
 import ExpenseDate from "../ExpenseDate/ExpenseDate";
 import "./ExpenseItem.css"; //import with css extension
