+++ src/components/Feature/Expenses/Expenses.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Card from '../Card/Card';
+import Card from '../../UI/Card/Card';
 import ExpenseItem from '../ExpenseItem/ExpenseItem';
 import "./Expenses.css"
